requirements.txt
manimce
networkx[all]
